[PORT FROM MAIN][DO NOT UPSTREAM]MTP server: MTP Server Tombstone after unplug/plug USB cable during transfering file.
BZ: 107499

Rootcause:
When user unplug USB cable during transfering file from PC to DUT,
MTP server call unlink system function to remove dirty file, and then call implemenation
of MTP java to update object information. At same time, USB driver notify MTP service in
Java space with USB disconnect event. MTP service release instance of MTP server and
MTP database. If unlink function block MTP server, MTP server will tombstone due to illegal access.
Solution:
Move unlink call for EIO error case.

Orig-Change-Id: Ie3a8037ef85f4ccbd2b890a453f112788a73047b
Category: aosp improvement
Domain: USB MTP-PTP
Origin: internal
Upstream-Candidate: no, proprietary

+++ media/mtp/MtpServer.cpp
@@ -235,6 +235,11 @@ void MtpServer::run() {
         mDatabase->sessionEnded();
     close(fd);
     mFD = -1;
+
+    if (mDirtyFilePath.length() != 0) {
+        ALOGD("Delete dirty file %s. \n", (const char *)mDirtyFilePath);
+        unlink(mDirtyFilePath);
+    }
 }
 
 void MtpServer::sendObjectAdded(MtpObjectHandle handle) {
@@ -959,7 +964,12 @@ MtpResponseCode MtpServer::doSendObject() {
     close(mfr.fd);
 
     if (ret < 0) {
-        unlink(mSendObjectFilePath);
+        if (errno == EIO) {
+            mDirtyFilePath = mSendObjectFilePath;
+        } else {
+            unlink(mSendObjectFilePath);
+        }
+
         if (errno == ECANCELED || errno == EIO )
             result = MTP_RESPONSE_TRANSACTION_CANCELLED;
         else
